test: move integration fixture into conftest
